# Conflicts:
#	services/src/pce/src/main/java/org/openkilda/pce/model/AvailableNetwork.java

+++ services/src/atdd-staging/src/main/java/org/openkilda/atdd/staging/steps/FlowCrudSteps.java
@@ -518,7 +518,7 @@ public class FlowCrudSteps implements En {
     @When("^change bandwidth of (.*) flow to '(.*)'$")
     public void changeBandwidthOfFlow(String flowAlias, String bwAlias) {
         FlowPayload flow = topologyUnderTest.getAliasedObject(flowAlias);
-        int bw = ((Long) topologyUnderTest.getAliasedObject(bwAlias)).intValue();
+        long bw = topologyUnderTest.getAliasedObject(bwAlias);
         flow.setMaximumBandwidth(bw);
     }
